# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for the GoChat",
        "title": "GoChat API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/webhook/storage/events": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Storage event",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Incorrect ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "basic"
        }
    }
}